// lib/api/generated/models/exerciseFamily.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export type ExerciseFamily = typeof ExerciseFamily[keyof typeof ExerciseFamily];


export const ExerciseFamily = {
  olympic: 'olympic',
  squat: 'squat',
  pull: 'pull',
  accessory: 'accessory',
  warmup: 'warmup',
} as const;